+++ src/PullStream.Csv/LastNewLine.cs
@@ -1,8 +1,0 @@
-﻿namespace PullStream.Csv
-{
-    public enum LastNewLine : sbyte
-    {
-        No,
-        Yes
-    }
-}