+++ CHANGELOG.md
@@ -7,8 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- NotDisposing [#41](https://github.com/pepelev/PullStream/issues/41)
+
+## [1.2.1] - 2021-06-15
+### Changed
+- Remove csv related code from pullstream branch [#47](https://github.com/pepelev/PullStream/issues/47)
+
+### Fixed
+- Validate all arguments of public methods [#49](https://github.com/pepelev/PullStream/issues/49)
 
 ## [1.2.0] - 2021-05-03
 ### Added
@@ -48,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic functionality #1 #3 #4 #6 #9 #10 #11 #12
 - Logo
 
-[Unreleased]: https://github.com/pepelev/pullstream/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/pepelev/pullstream/compare/v1.2.1...HEAD
+[1.2.1]: https://github.com/pepelev/pullstream/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/pepelev/pullstream/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/pepelev/pullstream/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/pepelev/pullstream/compare/v1.0.0...v1.0.1